% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgplot.R
\name{sg_vbar}
\alias{sg_vbar}
\title{Request a Vertical Bar Plot}
\usage{
sg_vbar(
  plt,
  var,
  response = NULL,
  stat = NULL,
  group = NULL,
  groupdisplay = NULL
)
}
\arguments{
\item{plt}{The base plotting object.  Value values are an \code{sgplot}
or \code{sgpanel} object.}

\item{var}{The variable to use for the X axis.}

\item{response}{The variable to use for the Y axis.}

\item{stat}{The statistic to use for the bar height calculation. This parameter}

\item{group}{The variable to group by.}

\item{groupdisplay}{Indicates how to display the grouping variable.
is optional.  If not supplied, the raw value of the response variable will
be used.}
}
\description{
Request a Vertical Bar Plot
}
