% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgplot.R
\name{sgpanel}
\alias{sgpanel}
\title{Initializes an SGPanel Object}
\usage{
sgpanel(dat, titles = NULL, theme = NULL)
}
\arguments{
\item{dat}{The data to use for this plot. This parameter is required.}

\item{titles}{A vector of optional titles.}

\item{theme}{An optional theme to use for this plot.}
}
\description{
Initializes an SGPanel Object
}
