% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgplot.R
\name{sgplot}
\alias{sgplot}
\title{Initializes an SGPlot Object}
\usage{
sgplot(dat, titles = NULL, theme = NULL)
}
\arguments{
\item{dat}{The data to use for this plot. This parameter is required.}

\item{titles}{A vector of optional titles.}

\item{theme}{An optional theme to use for this plot.}
}
\description{
The \code{sgplot} function creates an object of class "sgplot",
which forms the basis for a new plot.
}
