% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrender.R
\name{sgrender}
\alias{sgrender}
\title{Executes Rendering of a SGPlot Object}
\usage{
sgrender(
  plt,
  output_type = NULL,
  file_path = NULL,
  height = NULL,
  width = NULL,
  units = "px"
)
}
\arguments{
\item{plt}{The plot object to render. This parameter is required.}

\item{output_type}{An image type to output the plot to. Valid values
are "JPG", "JPEG", "SVG", "PNG", "BMP", "TIF", "TIFF" or NULL.  When NULL,
the function will render the plot interactively to the viewer.  Default is NULL.}

\item{file_path}{A path and file name to render the object to, if
an output file is requested. If no path is supplied,
a temporary file name will be generated.}

\item{height}{The height to of the output image.  Defaults to 480 pixels,
or the equivalent in the requested units.}

\item{width}{The width of the output image.  Defaults to 640 pixels, or the
equivalent in the requested units.}

\item{units}{The units associated with the image height and width. Defaults to
"px" meaning pixels.  Additional valid values are "in" for inches and "cm" for
centimeters.}
}
\value{
If an output file is requested, the path to the file. Otherwise,
a NULL.
}
\description{
Renders a plot to a file or the viewer.
}
