% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleattrs.R
\name{styleattrs}
\alias{styleattrs}
\title{Assigns Style Attributes for a Plot}
\usage{
styleattrs(
  plt,
  backcolor = NULL,
  wallcolor = NULL,
  datacolors = NULL,
  datacontrastcolors = NULL,
  datafillpatterns,
  datalinepatterns = NULL,
  datasymbols = NULL
)
}
\arguments{
\item{plt}{The plot object to modify the style attributes for.}

\item{backcolor}{The color to use for the background.}

\item{wallcolor}{The color to use for the plot background.}

\item{datacolors}{The colors to use for the data.}

\item{datacontrastcolors}{The colors to use for the data contrast.}

\item{datafillpatterns}{The patterns to fill the data.}

\item{datalinepatterns}{The patterns to draw the lines.}

\item{datasymbols}{The symbols to use for points.}
}
\value{
The plot to which the style attributes are assigned.
}
\description{
The \code{styleattr} function allows you to override the template style settings
for a particular plot.  The function contains the most common style parameters
you may want to specify.
}
\details{
Additional details...
}
